feat(Pagination)[OK-159] Create more generic Pagination and integrate it with BoardsPagination and SlidePagination
